more restructuring of check-template
Great deal of technical debt paid back on check-templates.rkt.  There's a cross product of template checking that is:

where it comes from: sits in dd, in @template, in an actual function definition

what it is compared against: type expression, the one in a named DD, an actual body expression

The overall function structure of that is mostly cleaned up now, and the proper layering of syntax and functions is in place in check-templates.rkt, but not yet templates.rkt:

grade-...   does quoting
grade-...*  picks up values from context (surrounding htdd or htdf)
check-...   is given values to operate on and does the actual checking of the rubric items

There's more work to do to make check-template/type-expr smarter, to try and more some of the load off the ad-hoc checking in templates.